+++ src/scene/ChangeLevel.cpp
@@ -1879,7 +1879,7 @@ long ARX_CHANGELEVEL_Pop_Index(ARX_CHANGELEVEL_INDEX * asi, long num)
 		return -1;
 	}
 
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 		compressed[i] = ~compressed[i];
 
 	long ssize = size;
@@ -1944,7 +1944,7 @@ long ARX_CHANGELEVEL_Pop_Zones_n_Lights(ARX_CHANGELEVEL_INDEX * asi, long num)
 		return -1;
 	}
 
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 		compressed[i] = ~compressed[i];
 
 	long ssize = size;
@@ -2109,7 +2109,7 @@ long ARX_CHANGELEVEL_Pop_Player(ARX_CHANGELEVEL_INDEX * asi, ARX_CHANGELEVEL_PLA
 
 	_pSaveBlock->Read(loadfile, compressed);
 
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 		compressed[i] = ~compressed[i];
 
 	long ssize = size;
@@ -2397,7 +2397,7 @@ long ARX_CHANGELEVEL_Pop_IO( const std::string& ident)
 		return -1;
 	}
 
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 		compressed[i] = ~compressed[i];
 
 	long ssize = size;
@@ -3451,7 +3451,7 @@ long ARX_CHANGELEVEL_Pop_Globals()
 		return -1;
 	}
 
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 	{
 		compressed[i] = ~compressed[i];
 	}
@@ -4167,7 +4167,7 @@ long ARX_CHANGELEVEL_Get_Player_LevelData(ARX_CHANGELEVEL_PLAYER_LEVEL_DATA * pl
 	}
 
 	// Un-Crypt
-	for (long i = 0; i < size; i += 2)
+	for (size_t i = 0; i < size; i += 2)
 		compressed[i] = ~compressed[i];
 
 	// Explode File
